Reworked payment + tax return to handle payments

+++ tax-tracker/src/main/java/com/skillstorm/taxtracker/dtos/TaxReturnDTO.java
@@ -1,11 +1,12 @@
 package com.skillstorm.taxtracker.dtos;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import com.skillstorm.taxtracker.models.Client;
 import com.skillstorm.taxtracker.models.Cpa;
 import com.skillstorm.taxtracker.models.EmploymentSector;
 
-public record TaxReturnDTO(Client client, Cpa cpa, int year, String status, int amountOwed, int amountPaid, int cost,
+public record TaxReturnDTO(Client client, Cpa cpa, int year, String status, BigDecimal amountOwed, BigDecimal amountPaid, BigDecimal cost,
 		LocalDate creationDate, LocalDate updateDate, EmploymentSector employmentSector) {
 
 }
